# server/docs/swagger.yaml
definitions:
  article.Article:
    properties:
      auth_id:
        description: 用户id
        type: string
      author:
        $ref: '#/definitions/system.SysUser'
        description: 作者
      categories:
        description: 分类
        items:
          $ref: '#/definitions/article.Category'
        type: array
      categories_id:
        description: 分类id
        items:
          type: integer
        type: array
      comment_count:
        description: 评论数
        type: integer
      create_time:
        type: string
      delete_time:
        $ref: '#/definitions/gorm.DeletedAt'
      disable_comments:
        description: 禁止评论
        type: boolean
      editor_type:
        description: 编辑器类型
        type: string
      html_content:
        description: html内容
        type: string
      id:
        type: string
      is_top:
        description: 是否置顶
        type: boolean
      likes:
        description: 点赞数
        type: integer
      md_content:
        description: markdown内容
        type: string
      meta_description:
        description: 头部描述
        type: string
      meta_key_words:
        description: 头部关键字
        type: string
      password:
        description: 访问密码
        type: string
      slug:
        description: 别名
        type: string
      status:
        description: 状态 DRAFT, PUBLISHED
        type: string
      summary:
        description: 摘要
        type: string
      tags:
        description: tags
        items:
          $ref: '#/definitions/article.Tag'
        type: array
      tags_id:
        description: tags id
        items:
          type: integer
        type: array
      thumbnail:
        description: 缩略图
        type: string
      title:
        description: 标题
        type: string
      update_time:
        type: string
      visits:
        description: 访问量
        type: integer
      word_count:
        description: 字数
        type: integer
    required:
    - status
    - title
    type: object
  article.Category:
    properties:
      children:
        description: 子类
        items:
          $ref: '#/definitions/article.Category'
        type: array
      create_time:
        type: string
      delete_time:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        description: 描述
        type: string
      id:
        type: string
      name:
        description: 分类名
        type: string
      parent_id:
        description: 父级ID
        type: integer
      password:
        description: 访问密码
        type: string
      slug:
        description: 别名
        type: string
      thumbnail:
        description: 缩略图
        type: string
      update_time:
        type: string
    required:
    - name
    type: object
  article.Tag:
    properties:
      color:
        description: 颜色
        type: string
      create_time:
        type: string
      delete_time:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: string
      name:
        description: 标签名
        type: string
      slug:
        description: 别名
        type: string
      thumbnail:
        description: 缩略图
        type: string
      update_time:
        type: string
    required:
    - name
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  request.ArticleRequest:
    properties:
      auth_id:
        description: 用户id
        type: string
      categories_id:
        description: 分类id
        items:
          type: integer
        type: array
      comment_count:
        description: 评论数
        type: integer
      disable_comments:
        description: 禁止评论
        type: boolean
      editor_type:
        description: 编辑器类型
        type: string
      html_content:
        description: html内容
        type: string
      is_top:
        description: 是否置顶
        type: boolean
      md_content:
        description: markdown内容
        type: string
      meta_description:
        description: 头部描述
        type: string
      meta_key_words:
        description: 头部关键字
        type: string
      password:
        description: 访问密码
        type: string
      slug:
        description: 别名
        type: string
      status:
        description: 状态 DRAFT, PUBLISHED
        type: string
      summary:
        description: 摘要
        type: string
      tags_id:
        description: tags id
        items:
          type: integer
        type: array
      thumbnail:
        description: 缩略图
        type: string
      title:
        description: 标题
        type: string
      word_count:
        description: 字数
        type: integer
    required:
    - status
    - title
    type: object
  request.LoginRequest:
    properties:
      captcha:
        type: string
      captcha_id:
        type: string
      password:
        type: string
      user_name:
        type: string
    required:
    - captcha
    - captcha_id
    - password
    - user_name
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
      success:
        type: boolean
    type: object
  system.SysRole:
    properties:
      children:
        items:
          $ref: '#/definitions/system.SysRole'
        type: array
      create_time:
        type: string
      delete_time:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: string
      parent_id:
        description: 父角色ID
        type: string
      role_name:
        description: 角色名
        type: string
      update_time:
        type: string
    type: object
  system.SysUser:
    properties:
      avatar:
        type: string
      create_time:
        type: string
      delete_time:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      id:
        type: string
      nick_name:
        type: string
      password:
        type: string
      phone:
        type: string
      roles:
        description: 关联到角色表
        items:
          $ref: '#/definitions/system.SysRole'
        type: array
      update_time:
        type: string
      user_name:
        type: string
    type: object
  theme.Theme:
    properties:
      auth_id:
        description: 作者ID
        type: string
      author:
        $ref: '#/definitions/system.SysUser'
        description: 作者
      create_time:
        type: string
      delete_time:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        description: 描述
        type: string
      id:
        type: string
      is_public:
        description: 是否公开
        type: boolean
      name:
        description: 名字
        type: string
      theme:
        description: 主题
        type: string
      thumbnail:
        description: 简略图
        type: string
      update_time:
        type: string
    required:
    - theme
    type: object
info:
  contact: {}
  description: This is a sample Server pets
  title: Swagger Example API
  version: 0.0.1
paths:
  /article/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建文章
        in: body
        name: article
        required: true
        schema:
          $ref: '#/definitions/request.ArticleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"操作成功"}'
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建文章
      tags:
      - article
  /article/delete:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 参数
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"操作成功"}'
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除文章
      tags:
      - article
  /article/detail:
    get:
      consumes:
      - application/json
      parameters:
      - description: 参数
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":article.Article,"msg":"操作成功"}'
          schema:
            $ref: '#/definitions/response.Response'
      summary: 查询文章详情
      tags:
      - article
  /article/history:
    get:
      consumes:
      - application/json
      parameters:
      - description: 参数
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":[]article.Article,"msg":"操作成功"}'
          schema:
            $ref: '#/definitions/response.Response'
      summary: 查询文章历史记录
      tags:
      - article
  /article/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: 参数
        in: query
        name: id
        required: true
        type: string
      - description: 查询文章
        in: body
        name: article
        required: true
        schema:
          $ref: '#/definitions/request.ArticleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":[]article.Article,"msg":"操作成功"}'
          schema:
            $ref: '#/definitions/response.Response'
      summary: 查询文章
      tags:
      - article
  /article/update:
    put:
      consumes:
      - application/json
      parameters:
      - description: 修改文章
        in: body
        name: article
        required: true
        schema:
          $ref: '#/definitions/article.Article'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":article2.Article,"msg":"操作成功"}'
          schema:
            $ref: '#/definitions/response.Response'
      summary: 修改文章
      tags:
      - article
  /category/create:
    post:
      consumes:
      - application/json
      description: 创建Category
      parameters:
      - description: 创建Category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/article.Category'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":article.Category,"msg":"操作成功"}'
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建Category
      tags:
      - category
  /category/delete:
    delete:
      consumes:
      - application/json
      description: 删除Category
      parameters:
      - description: 删除Category
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":article.Category,"msg":"操作成功"}'
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除Category
      tags:
      - category
  /category/update:
    put:
      consumes:
      - application/json
      description: 修改Category
      parameters:
      - description: 创建Category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/article.Category'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":article.Category,"msg":"操作成功"}'
          schema:
            $ref: '#/definitions/response.Response'
      summary: 修改Category
      tags:
      - category
  /file/delete/{id}:
    delete:
      parameters:
      - description: 参数
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{""code"":200,""message"":""删除成功""}'
          schema:
            type: string
        "400":
          description: '{""code"":400,""message"":""删除失败""}'
          schema:
            type: string
        "500":
          description: '{""code"":500,""message"":""服务器错误""}'
          schema:
            type: string
      summary: 文件删除
      tags:
      - file
  /file/list:
    get:
      parameters:
      - description: 关键字
        in: query
        name: key_word
        type: string
      - in: query
        name: page
        required: true
        type: integer
      - in: query
        name: page_size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{""code"":200,""message"":""删除成功""}'
          schema:
            type: string
        "400":
          description: '{""code"":400,""message"":""删除失败""}'
          schema:
            type: string
        "500":
          description: '{""code"":500,""message"":""服务器错误""}'
          schema:
            type: string
      summary: 文件列表
      tags:
      - file
  /file/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: '{""code"":200,""message"":""上传成功""}'
          schema:
            type: string
        "400":
          description: '{""code"":400,""message"":""上传失败""}'
          schema:
            type: string
        "500":
          description: '{""code"":500,""message"":""服务器错误""}'
          schema:
            type: string
      summary: 文件上传
      tags:
      - file
  /public/base/login:
    post:
      parameters:
      - description: 用户名, 密码, 验证码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/system.SysUser'
                msg:
                  type: string
              type: object
      summary: 用户登录
      tags:
      - User
  /public/base/register:
    post:
      parameters:
      - description: 用户名, 密码, 验证码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.SysUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/system.SysUser'
                msg:
                  type: string
              type: object
      summary: 用户注册
      tags:
      - User
  /tag/create:
    post:
      consumes:
      - application/json
      description: 创建tag
      parameters:
      - description: 创建tag
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/article.Tag'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":article.Tag,"msg":"操作成功"}'
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建tag
      tags:
      - tag
  /tag/delete:
    delete:
      consumes:
      - application/json
      description: 删除tag
      parameters:
      - description: 删除tag
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":article.Tag,"msg":"操作成功"}'
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除tag
      tags:
      - tag
  /tag/list:
    get:
      consumes:
      - application/json
      description: 查询tag
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":[]article.Tag,"msg":"操作成功"}'
          schema:
            $ref: '#/definitions/response.Response'
      summary: 查询tag
      tags:
      - tag
  /theme/create:
    post:
      consumes:
      - application/json
      description: 主题创建
      parameters:
      - description: 主题
        in: body
        name: theme
        required: true
        schema:
          $ref: '#/definitions/theme.Theme'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":theme.ResponseTheme,"msg":"ok"}'
          schema:
            $ref: '#/definitions/response.Response'
      summary: 主题创建
      tags:
      - theme
  /theme/delete:
    delete:
      consumes:
      - application/json
      description: 主题删除
      parameters:
      - description: 主题id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":theme.Theme,"msg":"ok"}'
          schema:
            $ref: '#/definitions/response.Response'
      summary: 主题删除
      tags:
      - theme
  /theme/detail:
    get:
      consumes:
      - application/json
      description: 主题详情
      parameters:
      - description: 主题id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":theme.Theme,"msg":"ok"}'
          schema:
            $ref: '#/definitions/response.Response'
      summary: 主题详情
      tags:
      - theme
  /theme/list:
    get:
      consumes:
      - application/json
      description: 主题列表
      parameters:
      - description: 关键词
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":[]theme.ResponseTheme,"msg":"ok"}'
          schema:
            $ref: '#/definitions/response.Response'
      summary: 主题列表
      tags:
      - theme
  /theme/public:
    get:
      consumes:
      - application/json
      description: 查询公开主题
      parameters:
      - description: 关键词
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":[]theme.ResponseTheme,"msg":"ok"}'
          schema:
            $ref: '#/definitions/response.Response'
      summary: 查询公开主题
      tags:
      - theme
  /theme/update:
    put:
      consumes:
      - application/json
      description: 主题更新
      parameters:
      - description: 主题
        in: body
        name: theme
        required: true
        schema:
          $ref: '#/definitions/theme.Theme'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":theme.Theme,"msg":"ok"}'
          schema:
            $ref: '#/definitions/response.Response'
      summary: 主题更新
      tags:
      - theme
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"